Tien

+++ app/Http/Controllers/Admins/LibraryController.php
@@ -13,7 +13,6 @@ class LibraryController extends Controller
     {
         $data['cate_library_count'] = DB::table('cate_librarys')->count();
         $data['library_count'] = DB::table('librarys')->count();
-
         view()->share($data);
     }
 
@@ -56,11 +55,11 @@ class LibraryController extends Controller
 
         $this->validate($request, [
             'name' => 'required|min:10',
-            'summary'=>'required|max:100',
+            'summary'=>'required|max:200',
             'content' => 'required|min:100',
         ], [
             'name.min' => 'Tên không được ít hơn 10 kí tự',
-            'contentt.min' => 'Tên không được ít hơn 100 kí tự',
+            'content.min' => 'Tên không được ít hơn 100 kí tự',
             'name.required' => 'Tên bài viết không được để trống',
             'content.required' => 'Nội dung bài viết không được để trống',
         ]);
@@ -84,7 +83,7 @@ class LibraryController extends Controller
 
         DB::table('librarys')->insert([
             'name' => $request->name,
-            'slug' => Str::slug($request->name . "-" . time()),
+            'slug' => Str::slug($request->name . "-" . time()).'.html',
             'image' => $file_name,
             'summary'=>$request->summary,
             'content' => $request->content,
@@ -136,12 +135,12 @@ class LibraryController extends Controller
 
         $this->validate($request, [
             'name' => 'required|min:10',
-            'summary' => 'required|max:100',
-            'contentt' => 'required|min:100',
+            'summary' => 'required|max:200',
+            'content' => 'required|min:100',
         ], [
             'name.min' => 'Tên không được ít hơn 10 kí tự',
-            'summary.max' => 'Tóm tắt không được quá 100 kí tự',
-            'contentt.min' => 'Tên không được ít hơn 100 kí tự',
+            'summary.max' => 'Tóm tắt không được quá 200 kí tự',
+            'content.min' => 'Tên không được ít hơn 100 kí tự',
             'name.required' => 'Tên bài viết không được để trống',
             'content.required' => 'Nội dung bài viết không được để trống',
         ]);
@@ -157,7 +156,7 @@ class LibraryController extends Controller
             }
             $file->move('images/librarys/', $image);
             $file_name = $image;
-            if (file_exists('images/librarys/' . $image_update[0]) && $image_update[0] != '') {
+            if (file_exists('images/librarys/' . $image_update[0]) && $image_update[0] != '' && $image_update[0] != 'logo.png') {
                 unlink('images/librarys/' . $image_update[0]);
             }
 
@@ -167,7 +166,7 @@ class LibraryController extends Controller
 
         DB::table('librarys')->where('id', $id)->update([
             'name' => $request->name,
-            'slug' => Str::slug($request->name . "-" . time()),
+            'slug' => Str::slug($request->name . "-" . time()).'.html',
             'image' => $file_name,
             'summary' => $request->summary,
             'content' => $request->content,
@@ -211,7 +210,18 @@ class LibraryController extends Controller
         ]);
         return redirect()->back()->with('thanhcong', 'Thành công');
     }
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function cate_index()
+    {
 
+        $data['cate_librarys'] = DB::table('cate_librarys')->get();
+
+        return view('admin.pages.librarys.cate_index', $data);
+    }
     /**
      * Show the form for creating a new resource.
      *
@@ -219,8 +229,8 @@ class LibraryController extends Controller
      */
     public function cate_create()
     {
-        $data['cate_librarys'] = DB::table('cate_librarys')->orderByDesc('id')->get();
-        return view('admin.pages.librarys.cate_create', $data);
+
+        return view('admin.pages.librarys.cate_create');
     }
 
     /**
@@ -231,23 +241,100 @@ class LibraryController extends Controller
      */
     public function cate_store(Request $request)
     {
+
         $this->validate($request, [
-            'name' => 'min:5',
+            'name' => 'min:3',
         ], [
-            'name.min' => 'Tên không được ít hơn 10 kí tự',
+            'name.min' => 'Tên không được ít hơn 3 kí tự',
         ]);
 
+//Kiểm tra định dạng ảnh
+        if ($request->hasFile('image')) {
 
+            $file = $request->file('image');
+
+            $name = $file->getClientOriginalName();
+            $image = Str::random(7) . "_image_" . $name;
+            while (file_exists('images/cate-librarys/' . $image)) {
+                $image = Str::random(7) . "_image_" . $name;
+            }
+            $file->move('images/cate-librarys/', $image);
+            $file_name = $image;
+
+        } else {
+            $file_name = 'logo.png';
+        }
         DB::table('cate_librarys')->insert([
             'name' => $request->name,
             'slug' => Str::slug($request->name),
+            'icon' => $file_name,
             'status' => 1,
             'created_at' => now()
         ]);
 
-        return redirect()->back()->with('thongbao', 'Thành công');
+        return redirect()->back()->with('thongbao', 'Thêm thành công');
+    }
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int $id
+     * @return \Illuminate\Http\Response
+     */
+    public function cate_edit($id)
+    {
+
+        $data['cate_library'] = DB::table('cate_librarys')->orderByDesc('id')->first();
+
+        return view('admin.pages.librarys.cate_edit', $data);
     }
 
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
+     * @return \Illuminate\Http\Response
+     */
+    public function cate_update(Request $request, $id)
+    {
+
+        $image_update = DB::table('cate_librarys')->where('id', '=', $id)->pluck('icon');
+
+        $this->validate($request, [
+            'name' => 'min:3',
+        ], [
+            'name.min' => 'Tên không được ít hơn 3 kí tự',
+        ]);
+
+        if ($request->hasFile('image')) {
+
+            $file = $request->file('image');
+
+            $name = $file->getClientOriginalName();
+            $image = Str::random(4) . "_image_" . $name;
+            while (file_exists('images/cate-librarys/' . $image)) {
+                $image = Str::random(4) . "_image_" . $name;
+            }
+            $file->move('images/cate-librarys/', $image);
+            $file_name = $image;
+            if (file_exists('images/cate-librarys/' . $image_update[0]) && $image_update[0] != ''&& $image_update[0] != 'logo.png') {
+                unlink('images/cate-librarys/' . $image_update[0]);
+            }
+
+        } else {
+            $file_name = DB::table('cate_librarys')->where('id', $id)->pluck('icon')->first();
+        }
+
+        DB::table('cate_librarys')->where('id', $id)->update([
+            'name' => $request->name,
+            'slug' => Str::slug($request->name),
+            'icon' => $file_name,
+            'status' => 1,
+            'created_at' => now()
+        ]);
+
+        return redirect()->route('cate_library.index')->with('thongbao', 'Sửa thành công');
+    }
     /**
      * Remove the specified resource from storage.
      *
